dream-fortune-ai/backend/requirements.txt
fastapi
uvicorn
gunicorn
google-generativeai
python-dotenv
pydantic
